+++ scratch/impls.py
@@ -24,23 +24,23 @@ MODEL_CFGS = {
 
 def bmm(top_scores: Tensor, routed_output: Tensor) -> Tensor:
     """Combine via batched matrix multiply: (T, 1, K) @ (T, K, D) -> (T, 1, D)."""
-    return torch.bmm(top_scores.unsqueeze(1), routed_output).squeeze(1)
+    return torch.bmm(top_scores.unsqueeze(1), routed_output.float()).squeeze(1)
 
 
 def bcast_sum(top_scores: Tensor, routed_output: Tensor) -> Tensor:
     """Combine via bcast multiply and sum: (T, K, 1) * (T, K, D) -> sum -> (T, D)."""
-    return (top_scores.unsqueeze(-1) * routed_output).sum(dim=1)
+    return (top_scores.unsqueeze(-1) * routed_output.float()).sum(dim=1)
 
 
 @torch.compile
 def bcast_sum_compiled(top_scores: Tensor, routed_output: Tensor) -> Tensor:
     """Combine via bcast multiply and sum: (T, K, 1) * (T, K, D) -> sum -> (T, D)."""
-    return (top_scores.unsqueeze(-1) * routed_output).sum(dim=1)
+    return (top_scores.unsqueeze(-1) * routed_output.float()).sum(dim=1)
 
 
 def einsum(top_scores: Tensor, routed_output: Tensor) -> Tensor:
     """Combine via einsum: tk,tkd->td."""
-    return torch.einsum("tk,tkd->td", top_scores, routed_output)
+    return torch.einsum("tk,tkd->td", top_scores, routed_output.float())
 
 
 IMPLS: dict[str, Callable[[Tensor, Tensor], Tensor]] = {
